Revert "Add io/fs#FS Driver #471 (#472)"
This reverts commit fedbe42fc81800abd63c2a256159a51f28f49d67.
