+++ source/file/file.go
@@ -1,22 +1,31 @@
 package file
 
 import (
+	"net/http"
 	nurl "net/url"
 	"os"
 	"path/filepath"
 
 	"github.com/golang-migrate/migrate/v4/source"
+	"github.com/golang-migrate/migrate/v4/source/httpfs"
 )
 
 func init() {
 	source.Register("file", &File{})
 }
 
-func parseURL(url string) (string, error) {
+type File struct {
+	httpfs.PartialDriver
+	url  string
+	path string
+}
+
+func (f *File) Open(url string) (source.Driver, error) {
 	u, err := nurl.Parse(url)
 	if err != nil {
-		return "", err
+		return nil, err
 	}
+
 	// concat host and path to restore full path
 	// host might be `.`
 	p := u.Opaque
@@ -28,7 +37,7 @@ func parseURL(url string) (string, error) {
 		// default to current directory if no path
 		wd, err := os.Getwd()
 		if err != nil {
-			return "", err
+			return nil, err
 		}
 		p = wd
 
@@ -36,9 +45,17 @@ func parseURL(url string) (string, error) {
 		// make path absolute if relative
 		abs, err := filepath.Abs(p)
 		if err != nil {
-			return "", err
+			return nil, err
 		}
 		p = abs
 	}
-	return p, nil
+
+	nf := &File{
+		url:  url,
+		path: p,
+	}
+	if err := nf.Init(http.Dir(p), ""); err != nil {
+		return nil, err
+	}
+	return nf, nil
 }
