--- source/iofs/README.md
@@ -1,3 +1,0 @@
-# iofs
-
-https://pkg.go.dev/github.com/golang-migrate/migrate/v4/source/iofs
